+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), ~~and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)~~.
 
+## [Unreleased]
+### Fixed
+ - Fix bug to handle default language correctly
+
 ## [1.4.0] - 2018-11-27
 ### Added
  - Add document language filter ([#20](https://github.com/pwyf/iati-decipher/issues/20))